requirements.txt
faiss-cpu
sentence-transformers
numpy
requests
python-dotenv
pypdf
openai